updated verification script to run only in case a updated configuration file has been provided

+++ openssl/verify.sh
@@ -6,33 +6,40 @@ filesToProcess() {
 }
 
 cd openssl
-mkdir --p "$1"
 
-#Ref first
-i=0
+if [ -f configFlags ]
+  then
+    
+	mkdir --p "$1"
 
-filesToProcess|while read feature; do
-        echo "Verifiying $1"
-        echo "With features: $feature"
-        ./genConfig.sh $feature | tee "$1"/"$i"_ref.config
-        ./build.sh | tee "$1"/"$i"_ref.build
-        ./runtest.sh | tee "$1"/"$i"_ref.test
-        i=`expr $i + 1`
+	#Ref first
+	i=0
+
+	filesToProcess|while read feature; do
+        	echo "Verifiying $1"
+        	echo "With features: $feature"
+        	./genConfig.sh $feature | tee "$1"/"$i"_ref.config
+        	./build.sh | tee "$1"/"$i"_ref.build
+        	./runtest.sh | tee "$1"/"$i"_ref.test
+        	i=`expr $i + 1`
         done
 
-#Clean
-git checkout .
+	#Clean
+	git checkout .
 
-#Org Next
-i=0
+	#Org Next
+	i=0
 
-filesToProcess|while read feature; do
- 	    echo "Verifiying $1"
-        echo "With features: $feature"
-        ./genConfig.sh $feature | tee "$1"/"$i"_org.config
-        ./build.sh | tee "$1"/"$i"_org.build
+	filesToProcess|while read feature; do
+ 		echo "Verifiying $1"
+        	echo "With features: $feature"
+        	./genConfig.sh $feature | tee "$1"/"$i"_org.config
+        	./build.sh | tee "$1"/"$i"_org.build
  	 	./runtest.sh | tee "$1"/"$i"_org.test
-        i=`expr $i + 1`
+        	i=`expr $i + 1`
         done
 
+	rm configFlags
+
+fi
 exit
